LayoutTests/media/modern-media-controls/airplay-button/airplay-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>AirplayButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const airplayButton = new AirplayButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("airplayButton.element.localName", "button");
shouldBeTrue('airplayButton.element.classList.contains("icon")');
shouldBeTrue('airplayButton.element.classList.contains("airplay")');
shouldBeEqualToString("airplayButton.iconName", `${Icons.Airplay}`);

const macOSFullscreenAirplayButton = new AirplayButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });

const iOSAirplayButton = new AirplayButton({ layoutTraits: LayoutTraits.iOS });

scheduler.frameDidFire = function()
{
    if ([airplayButton, macOSFullscreenAirplayButton, iOSAirplayButton].some(button => button.width == 0))
        return;

    shouldBeTrue('airplayButton.element.style.webkitMaskImage.includes("macOS/airplay@")');
    shouldBeTrue('macOSFullscreenAirplayButton.element.style.webkitMaskImage.includes("macOS/airplay-fullscreen@")');
    shouldBeTrue('iOSAirplayButton.element.style.webkitMaskImage.includes("iOS/airplay@")');

    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
